backend/telegram_scraper/requirements.txt
amqp==5.2.0
asgiref==3.7.2
attrs==24.2.0
Authlib==1.3.1
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
configparser==7.1.0
cron_descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.8.0rc2
dj-database-url==2.1.0
dj-rest-auth==6.0.0
dj-static==0.0.6
Django==4.2.9
django-allauth==64.0.0
django-bootstrap-v5==1.0.11
django-celery-beat==2.5.0
django-ckeditor==6.7.1
django-cors-headers==4.3.1
django-heroku==0.3.1
django-js-asset==2.2.0
django-microsoft-auth==3.0.1
django-microsoft-authentication==0.1.6
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
dnspython==2.6.1
et-xmlfile==1.1.0
flower==2.0.1
gunicorn==20.0.4
h11==0.14.0
humanize==4.9.0
idna==3.6
kombu==5.3.4
lxml==5.3.0
msal==1.26.0
numpy==2.1.1
oauthlib==3.2.2
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.1
pandas==2.2.2
pillow==10.2.0
prometheus-client==0.8.0
prompt-toolkit==3.0.43
psycopg2==2.9.9
psycopg2-binary==2.9.9
pyaes==1.6.1
pyasn1==0.6.1
pycparser==2.21
PyJWT==2.8.0
pymongo==4.8.0
pypi==2.1
PySocks==1.7.1
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-oauth2==1.1.1
python-social-auth==0.3.6
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rest-framework-simplejwt==0.0.2
rest_social_auth==8.4.0
rsa==4.9
selenium==4.24.0
sentry-sdk==1.39.1
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.0
social-auth-core==4.5.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
static3==0.7.0
Telethon==1.36.0
tornado==6.4
trio==0.26.2
trio-websocket==0.11.1
typing_extensions==4.10.0
tzdata==2023.4
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
websocket-client==1.8.0
whitenoise==6.6.0
wsproto==1.2.0
